requirements.txt
Django>3.0.5,<3.1.0
djangorestframework>3.10.3,<3.12.0
Pillow>7.1.1,<7.2.0
psycopg2>=2.8.5,<2.9.0
